requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
google-genai==0.3.0
pypdf==5.1.0
python-docx==1.1.2
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
jinja2==3.1.6
